third_party/WebKit/LayoutTests/fast/mediastream/MediaStream-clone-expected.txt

Test cloning mediastreams and their tracks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS stream.getAudioTracks().length is 1
PASS stream.getVideoTracks().length is 1
PASS stream.id is not cloned_stream.id
PASS cloned_stream.getAudioTracks().length is 1
PASS cloned_stream.getVideoTracks().length is 1
PASS stream.getAudioTracks()[0] is not cloned_stream.getAudioTracks()[0]
PASS stream.getVideoTracks()[0] is not cloned_stream.getVideoTracks()[0]
PASS stream.getAudioTracks()[0].id is not cloned_stream.getAudioTracks()[0].id
PASS stream.getVideoTracks()[0].id is not cloned_stream.getVideoTracks()[0].id
PASS track is not cloned_track
PASS track.id is not cloned_track.id
PASS stream.getAudioTracks().length is 2
PASS successfullyParsed is true

TEST COMPLETE
